Cross cutting concerns & Global usings

+++ EshopMicroservices/src/Services/Catalog.API/Catalog.API/Products/DeleteProductById/DeleteProductByIdCommandHandler.cs
@@ -1,18 +1,24 @@
-﻿using BuildingBlocks.CQRS;
-using Catalog.API.Models;
-using Marten;
-using MediatR;
-using System.Windows.Input;
-using ICommand = BuildingBlocks.CQRS.ICommand;
-
-namespace Catalog.API.Products.DeleteProductById
+﻿namespace Catalog.API.Products.DeleteProductById
 {
     public record DeleteProductByIdCommand(Guid Id) : ICommand<DeleteProductByIdResult>;
     public record DeleteProductByIdResult(bool IsSuccess);
+
+    public class DeleteProductValidator : AbstractValidator<DeleteProductByIdCommand>
+    {
+        public DeleteProductValidator()
+        {
+            RuleFor(x => x.Id).NotEmpty().WithMessage("Id cannot be empty");
+        }
+    }
     public class DeleteProductByIdCommandHandler(IDocumentSession session, ILogger<DeleteProductByIdCommandHandler> logger) : ICommandHandler<DeleteProductByIdCommand, DeleteProductByIdResult>
     {
         public async Task<DeleteProductByIdResult> Handle(DeleteProductByIdCommand command, CancellationToken cancellationToken)
         {
+            var product = await session.LoadAsync<Product>(command.Id, cancellationToken);
+            if (product == null)
+            {
+                throw new ProductNotFoundException(command.Id);
+            }
             session.Delete<Product>(command.Id);
             await session.SaveChangesAsync(cancellationToken);
             return new DeleteProductByIdResult(true);
